+++ app/src/main/java/lijewski/samplerestapp/SampleRestApp.kt
@@ -1,0 +1,19 @@
+package lijewski.samplerestapp
+
+import dagger.android.AndroidInjector
+import dagger.android.DaggerApplication
+import dagger.android.HasAndroidInjector
+import lijewski.samplerestapp.di.app.DaggerAppComponent
+import timber.log.Timber
+
+class SampleRestApp : DaggerApplication(), HasAndroidInjector {
+    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
+        return DaggerAppComponent.factory().create(this)
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+
+        Timber.plant(Timber.DebugTree())
+    }
+}
