+++ app/src/main/java/lijewski/samplerestapp/databinding/CustomBindingAdapter.kt
@@ -1,0 +1,23 @@
+package lijewski.samplerestapp.databinding
+
+import android.view.View
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+
+@BindingAdapter("visibleGone")
+fun visibleGone(view: View, show: Boolean) {
+    view.visibility = if (show) View.VISIBLE else View.GONE
+}
+
+@BindingAdapter("visibleInvisible")
+fun visibleInvisible(view: View, show: Boolean) {
+    view.visibility = if (show) View.VISIBLE else View.INVISIBLE
+}
+
+@BindingAdapter("loadImage")
+fun loadImage(view: ImageView, imageUrl: String) {
+    Glide.with(view.context)
+        .load(imageUrl)
+        .into(view)
+}
