+++ app/src/main/java/lijewski/samplerestapp/di/module/AppModule.kt
@@ -1,0 +1,14 @@
+package lijewski.samplerestapp.di.module
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import lijewski.samplerestapp.SampleRestApp
+
+@Module(includes = [ViewModelModule::class])
+class AppModule {
+    @Provides
+    fun provideContext(application: SampleRestApp): Context {
+        return application.applicationContext
+    }
+}
