+++ app/src/main/java/lijewski/samplerestapp/di/module/ViewModelModule.kt
@@ -1,0 +1,22 @@
+package lijewski.samplerestapp.di.module
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import dagger.Binds
+import dagger.Module
+import dagger.multibindings.IntoMap
+import lijewski.samplerestapp.di.key.ViewModelKey
+import lijewski.samplerestapp.presentation.factory.ViewModelFactory
+import lijewski.samplerestapp.presentation.main.dashboard.DashboardViewModel
+
+@Suppress("unused")
+@Module
+abstract class ViewModelModule {
+    @Binds
+    abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
+
+    @Binds
+    @IntoMap
+    @ViewModelKey(DashboardViewModel::class)
+    abstract fun bindDashboardViewModel(dashboardViewModel: DashboardViewModel): ViewModel
+}
