--- app/src/main/java/lijewski/songapp/di/module/RepositoryModule.kt
@@ -1,18 +1,0 @@
-package lijewski.songapp.di.module
-
-import dagger.Module
-import dagger.Provides
-import lijewski.data.network.RemoteApi
-import lijewski.data.repository.RemoteRepositoryImpl
-import lijewski.data.mapper.ResultMapper
-import lijewski.domain.repository.RemoteRepository
-import javax.inject.Singleton
-
-@Module
-class RepositoryModule {
-    @Provides
-    @Singleton
-    fun provideSongRepository(remoteApi: RemoteApi, resultMapper: ResultMapper): RemoteRepository {
-        return RemoteRepositoryImpl(remoteApi, resultMapper)
-    }
-}
