--- app/src/main/java/lijewski/songapp/di/module/ViewModelModule.kt
@@ -1,22 +1,0 @@
-package lijewski.songapp.di.module
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import dagger.Binds
-import dagger.Module
-import dagger.multibindings.IntoMap
-import lijewski.songapp.di.key.ViewModelKey
-import lijewski.songapp.presentation.factory.ViewModelFactory
-import lijewski.songapp.presentation.main.dashboard.DashboardViewModel
-
-@Suppress("unused")
-@Module
-abstract class ViewModelModule {
-    @Binds
-    abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(DashboardViewModel::class)
-    abstract fun bindDashboardViewModel(dashboardViewModel: DashboardViewModel): ViewModel
-}
