+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">SongApp</string>
+    <string name="app_name">SampleRestApp</string>
     <string name="action_settings">Settings</string>
     <string name="no_songs">No songs found</string>
     <string name="error_fetch">Failed to fetch searchResult list</string>
